refactor(#260): consolidate type-only modules into types.ts files
