+++ projects/app-web/src/lib/auth/get-verify-session.ts
@@ -1,6 +1,6 @@
 import { getSession } from '@tanstack/react-start/server';
 import { buildVerifySessionConfig } from './build-verify-session-config';
-import type { VerifySessionData } from './verify-session-data';
+import type { VerifySessionData } from './types';
 
 /** Reads the caller's in-flight verification state; empty once the 10-minute window lapses. */
 export async function getVerifySession(): Promise<VerifySessionData> {
