+++ projects/app-web/src/lib/auth/types.ts
@@ -1,3 +1,11 @@
+/** The `en_session` cookie's stored shape: the caller's live session plus its target expiry. */
+export interface AuthSessionData {
+  readonly accessToken?: string;
+  readonly expires?: string;
+  readonly refreshToken?: string;
+  readonly sessionID?: string;
+}
+
 /**
  * Keys are prefixed with the flow they belong to so two in-flight flows sharing the one
  * `en_verification` cookie can't clobber each other's state.
