+++ projects/app-web/src/lib/auth/update-auth-session.ts
@@ -1,9 +1,9 @@
 import { getSession, updateSession } from '@tanstack/react-start/server';
-import type { AuthSessionData } from './auth-session-data';
 import {
   AUTH_SESSION_READ_MAX_AGE_SECONDS,
   buildAuthSessionConfig,
 } from './build-auth-session-config';
+import type { AuthSessionData } from './types';
 
 export interface UpdateAuthSessionOptions {
   readonly expiresAt?: Date;
