+++ projects/app-web/src/routes/-verify-otp/run-change-email.ts
@@ -1,6 +1,6 @@
 import { redirect } from '@tanstack/react-router';
 import { userClient } from '../../lib/rpc/clients/user-client';
-import type { RunVerificationContext } from './run-verification-context';
+import type { RunVerificationContext } from './types';
 
 /** Confirms ownership of a new email address and applies it to the caller's own account. */
 export async function runChangeEmail(ctx: Readonly<RunVerificationContext>): Promise<never> {
