+++ projects/app-web/src/routes/-verify-otp/run-onboarding.ts
@@ -1,6 +1,6 @@
 import { redirect } from '@tanstack/react-router';
 import { updateVerifySession } from '../../lib/auth/update-verify-session';
-import type { RunVerificationContext } from './run-verification-context';
+import type { RunVerificationContext } from './types';
 
 /** Records the verified email so the onboarding page's guard can trust it, then sends the caller there. */
 export async function runOnboarding(ctx: Readonly<RunVerificationContext>): Promise<never> {
