--- projects/app-web/src/routes/-verify-otp/run-verification-context.ts
@@ -1,5 +1,0 @@
-/** What a verification type's post-verify continuation needs: who/what it verified for. */
-export interface RunVerificationContext {
-  readonly redirectTo?: string | undefined;
-  readonly target: string;
-}
