+++ projects/app-web/src/routes/-verify-otp/types.ts
@@ -1,3 +1,9 @@
+/** What a verification type's post-verify continuation needs: who/what it verified for. */
+export interface RunVerificationContext {
+  readonly redirectTo?: string | undefined;
+  readonly target: string;
+}
+
 /**
  * The verify-otp form's non-redirect outcome. Every successful verify ends in a thrown redirect
  * instead of a value here — its continuation depends on the verification type.
